feat: organize client session tests by domain
What changed and why:
Split monolithic test_request_handling.py into domain-specific test
files:
- test_request_routing.py (ping + routing logic)
- test_roots_handling.py (filesystem roots functionality)
- test_sampling_handling.py (LLM sampling requests)
- test_elicitation_handling.py (server elicitation requests)

Also organized request handlers in client session by domain with clear
section headers matching the test structure.

Impact:
- Improved test discoverability and maintainability
- Consistent organization pattern between client and server tests
- Easier parallel development on different client capabilities
- Clear separation of concerns for each domain
