+++ src/conduit/client/session.py
@@ -239,7 +239,113 @@ class ClientSession(BaseSession):
         self.server_state.info = result.server_info
 
     # ================================
-    # Request handlers
+    # Roots
+    # ================================
+
+    async def _handle_list_roots(
+        self, request: ListRootsRequest
+    ) -> ListRootsResult | Error:
+        """Handle server request for filesystem roots.
+
+        Only processes requests if the client advertised roots capability during
+        initialization. Delegates actual roots logic to the RootsManager.
+
+        Args:
+            request: Parsed roots/list request from server.
+
+        Returns:
+            ListRootsResult from the roots manager, or Error if:
+            - Client didn't advertise roots capability (METHOD_NOT_FOUND)
+            - Handler raised unexpected exception (INTERNAL_ERROR)
+        """
+        if self.client_config.capabilities.roots is None:
+            return Error(
+                code=METHOD_NOT_FOUND,
+                message="Client does not support roots capability",
+            )
+        try:
+            return await self.roots.handle_list_roots(request)
+        except Exception:
+            return Error(code=INTERNAL_ERROR, message="Error in roots handler")
+
+    # ================================
+    # Sampling
+    # ================================
+
+    async def _handle_sampling(
+        self, request: CreateMessageRequest
+    ) -> CreateMessageResult | Error:
+        """Handle server request for LLM sampling.
+
+        Only processes requests if the client advertised sampling capability during
+        initialization. Delegates actual sampling logic to the SamplingManager.
+
+        Args:
+            request: Parsed sampling/createMessage request from server.
+
+        Returns:
+            CreateMessageResult from the configured handler, or Error if:
+            - Client didn't advertise sampling capability (METHOD_NOT_FOUND)
+            - No sampling handler configured (METHOD_NOT_FOUND)
+            - Handler raised unexpected exception (INTERNAL_ERROR)
+        """
+        if not self.client_config.capabilities.sampling:
+            return Error(
+                code=METHOD_NOT_FOUND,
+                message="Client does not support sampling capability",
+            )
+        try:
+            return await self.sampling.handle_create_message(request)
+        except SamplingNotConfiguredError as e:
+            return Error(code=METHOD_NOT_FOUND, message=str(e))
+        except Exception:
+            return Error(code=INTERNAL_ERROR, message="Error in sampling handler")
+
+    # ================================
+    # Elicitation
+    # ================================
+
+    async def _handle_elicitation(self, request: ElicitRequest) -> ElicitResult | Error:
+        """Handle server request for elicitation.
+
+        Only processes requests if the client advertised elicitation capability during
+        initialization. Delegates actual elicitation logic to the ElicitationManager.
+
+        Args:
+            request: Parsed elicitation/create request from server.
+
+        Returns:
+            ElicitResult from the configured handler, or Error if:
+            - Client didn't advertise elicitation capability (METHOD_NOT_FOUND)
+            - No elicitation handler configured (METHOD_NOT_FOUND)
+            - Handler raised unexpected exception (INTERNAL_ERROR)
+        """
+        if not self.client_config.capabilities.elicitation:
+            return Error(
+                code=METHOD_NOT_FOUND,
+                message="Client does not support elicitation capability",
+            )
+        try:
+            return await self.elicitation.handle_elicitation(request)
+        except ElicitationNotConfiguredError as e:
+            return Error(code=METHOD_NOT_FOUND, message=str(e))
+        except Exception:
+            return Error(code=INTERNAL_ERROR, message="Error in elicitation handler")
+
+    # ================================
+    # Ping
+    # ================================
+
+    async def _handle_ping(self, request: PingRequest) -> EmptyResult | Error:
+        """Handle server request for ping.
+
+        Returns:
+            PingResult with pong.
+        """
+        return EmptyResult()
+
+    # ================================
+    # Request routing
     # ================================
 
     async def _handle_session_request(self, payload: dict[str, Any]) -> Result | Error:
@@ -284,96 +390,6 @@ class ClientSession(BaseSession):
             "sampling/createMessage": (CreateMessageRequest, self._handle_sampling),
             "elicitation/create": (ElicitRequest, self._handle_elicitation),
         }
-
-    async def _handle_ping(self, request: PingRequest) -> EmptyResult | Error:
-        """Handle server request for ping.
-
-        Returns:
-            PingResult with pong.
-        """
-        return EmptyResult()
-
-    async def _handle_list_roots(
-        self, request: ListRootsRequest
-    ) -> ListRootsResult | Error:
-        """Handle server request for filesystem roots.
-
-        Only processes requests if the client advertised roots capability during
-        initialization. Delegates actual roots logic to the RootsManager.
-
-        Args:
-            request: Parsed roots/list request from server.
-
-        Returns:
-            ListRootsResult from the roots manager, or Error if:
-            - Client didn't advertise roots capability (METHOD_NOT_FOUND)
-            - Handler raised unexpected exception (INTERNAL_ERROR)
-        """
-        if self.client_config.capabilities.roots is None:
-            return Error(
-                code=METHOD_NOT_FOUND,
-                message="Client does not support roots capability",
-            )
-        try:
-            return await self.roots.handle_list_roots(request)
-        except Exception:
-            return Error(code=INTERNAL_ERROR, message="Error in roots handler")
-
-    async def _handle_sampling(
-        self, request: CreateMessageRequest
-    ) -> CreateMessageResult | Error:
-        """Handle server request for LLM sampling.
-
-        Only processes requests if the client advertised sampling capability during
-        initialization. Delegates actual sampling logic to the SamplingManager.
-
-        Args:
-            request: Parsed sampling/createMessage request from server.
-
-        Returns:
-            CreateMessageResult from the configured handler, or Error if:
-            - Client didn't advertise sampling capability (METHOD_NOT_FOUND)
-            - No sampling handler configured (METHOD_NOT_FOUND)
-            - Handler raised unexpected exception (INTERNAL_ERROR)
-        """
-        if not self.client_config.capabilities.sampling:
-            return Error(
-                code=METHOD_NOT_FOUND,
-                message="Client does not support sampling capability",
-            )
-        try:
-            return await self.sampling.handle_create_message(request)
-        except SamplingNotConfiguredError as e:
-            return Error(code=METHOD_NOT_FOUND, message=str(e))
-        except Exception:
-            return Error(code=INTERNAL_ERROR, message="Error in sampling handler")
-
-    async def _handle_elicitation(self, request: ElicitRequest) -> ElicitResult | Error:
-        """Handle server request for elicitation.
-
-        Only processes requests if the client advertised elicitation capability during
-        initialization. Delegates actual elicitation logic to the ElicitationManager.
-
-        Args:
-            request: Parsed elicitation/create request from server.
-
-        Returns:
-            ElicitResult from the configured handler, or Error if:
-            - Client didn't advertise elicitation capability (METHOD_NOT_FOUND)
-            - No elicitation handler configured (METHOD_NOT_FOUND)
-            - Handler raised unexpected exception (INTERNAL_ERROR)
-        """
-        if not self.client_config.capabilities.elicitation:
-            return Error(
-                code=METHOD_NOT_FOUND,
-                message="Client does not support elicitation capability",
-            )
-        try:
-            return await self.elicitation.handle_elicitation(request)
-        except ElicitationNotConfiguredError as e:
-            return Error(code=METHOD_NOT_FOUND, message=str(e))
-        except Exception:
-            return Error(code=INTERNAL_ERROR, message="Error in elicitation handler")
 
     # ================================
     # Notification handlers
